LLM/LLaMA/requirement.txt
pytorch==2.3.1
transformers==4.48.1